src/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.2
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
executing==1.1.1
flake8==5.0.4
flake8-quotes==3.3.1
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
gunicorn==20.1.0
hiredis==2.0.0
idna==3.4
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
jsonfield==3.1.0
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-vcr==1.0.2
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
scipy==1.9.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
vcrpy==4.2.1
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.8.1
